Refactor: Telas cliente juridico
