+++ src/pages/Juridica/ClienteJuridico.module.css
@@ -6,16 +6,6 @@
   font-weight: bold;
   color: #393939;
 }
-
-.certificados {
-  width: 40%;
-}
-.certificadosTable {
-  width: 100%;
-}
-.adicionarButton {
-  width: 100%;
-}
 .funcionarios {
   width: 100%;
 }
