+++ src/pages/Juridica/ClienteJuridico.tsx
@@ -1,104 +1,31 @@
-import { Flex, Text, Title } from "@mantine/core";
-import { modals } from "@mantine/modals";
-import { showNotification } from "@mantine/notifications";
+import { Flex } from "@mantine/core";
 import { useQuery } from "@tanstack/react-query";
-import { AxiosError } from "axios";
-import { DataTable } from "mantine-datatable";
-import { useParams } from "react-router-dom";
+import { useNavigate, useParams } from "react-router-dom";
+import Certificados from "../../components/Certificado/Certificados";
 import EnderecoCliente from "../../components/EnderecoCliente";
-import ListaFuncionarioJuridico from "../../components/Funcionarios/ListaFuncionarioJuridico";
+import ListaFuncionarioJuridico from "../../components/Funcionarios/ListaFuncionario";
 import InfoClienteJ from "../../components/InfoClienteJ";
-import ModalCadCertificado from "../../components/ModalCadCertificado";
-import ModalEditCertificado from "../../components/ModalEditCertificado";
-import { Certificado } from "../../services/certificado/Certificado";
-import deletaCertificado from "../../services/certificado/deletaCertificado";
-import listaCertificadoById from "../../services/certificado/listaCertificadoById";
 import listaClienteById from "../../services/client/listaClienteById";
 import style from "./ClienteJuridico.module.css";
+import { useEffect } from "react";
 
 export default function ClienteJuridico() {
+  const navigate = useNavigate();
   const { id } = useParams() as { id: string };
 
-  const { data: Client } = useQuery({
-    queryKey: ["client"],
+  const {isError,  data: Client } = useQuery({
+    queryKey: ["clientJ"],
     queryFn: async () => listaClienteById(id),
   });
 
-  const {
-    data: certificados,
-    isFetching,
-    isRefetching,
-    refetch,
-  } = useQuery({
-    queryKey: ["certificados"],
-    queryFn: async () => listaCertificadoById(id),
-  });
-
-  const atualizaDados = () => {
-    refetch();
-  };
-
-  const AbreModalCadastrarCertificado = (certificados: Certificado) => {
-    modals.open({
-      title: "Cadastrar certificado",
-      children: (
-        <ModalCadCertificado
-          certificado={certificados}
-          close={modals.closeAll}
-          onSuccess={atualizaDados}
-        />
-      ),
-    });
-  };
-  const AbreModalEditCertificado = (certificados: Certificado) =>
-    modals.open({
-      title: "Editar certificado",
-      children: (
-        <ModalEditCertificado
-          certificado={certificados}
-          close={modals.closeAll}
-          onSuccess={atualizaDados}
-        />
-      ),
-    });
-
-  const excluirCertificado = (certificado: Certificado) =>
-    modals.openConfirmModal({
-      title: "Exclusão",
-      children: (
-        <Text size="sm">
-          Você está prestes a excluir o chip de número "{certificado.nome}". Tem
-          certeza disso?
-        </Text>
-      ),
-      labels: { confirm: "Confirmar Exclusão", cancel: "Cancelar" },
-      onConfirm: () => exclui(certificado._id),
-    });
-
-  const exclui = async (_id: string) => {
-    try {
-      await deletaCertificado(_id);
-      showNotification({
-        title: "Ok",
-        message: "Exclusão efetuada com sucesso!",
-        color: "green",
-      });
-      refetch().catch(() => {});
-    } catch (error) {
-      const message =
-        error instanceof AxiosError
-          ? error?.response?.data?.message
-          : "Ocorreu um erro ao excluir. Por favor, tente novamente";
-      showNotification({
-        title: "Erro",
-        message,
-        color: "red",
-      });
+  useEffect(() => {
+    if (isError) {
+      navigate("/juridica");
     }
-  };
+  }, [isError, id]);
 
   return (
-    <div>
+    <Flex direction='column'>
       <Flex
         mt="xs"
         wrap="wrap"
@@ -125,74 +52,11 @@ export default function ClienteJuridico() {
           Cep={Client?.cep}
           Pais={Client?.pais}
         />
-        <Flex direction="column" align="center" className={style.certificados}>
-          <Title order={2} mb="xs" className={style.titulo}>
-            Certificados
-          </Title>
-          <DataTable
-            minHeight={200}
-            withBorder
-            shadow="sm"
-            striped
-            highlightOnHover
-            horizontalSpacing="xl"
-            verticalAlignment="center"
-            className={style.certificadosTable}
-            fetching={isFetching || isRefetching}
-            records={certificados || []}
-            idAccessor="_id"
-            columns={[
-              { accessor: "nome", title: "Nome", textAlignment: "center" },
-              { accessor: "senha", title: "Senha", textAlignment: "center" },
-              {
-                accessor: "descricao",
-                title: "Descrição",
-                textAlignment: "center",
-              },
-              {
-                accessor: "dtvalidade",
-                title: "Validade",
-                textAlignment: "center",
-              },
-            ]}
-            noRecordsText="Nenhum registro encontrado!"
-            rowContextMenu={{
-              trigger: "click",
-              items: (record) => [
-                {
-                  key: "Cadastrar",
-                  onClick: () => AbreModalCadastrarCertificado(record),
-                },
-                {
-                  key: "editar",
-                  onClick: () => AbreModalEditCertificado(record),
-                },
-                {
-                  key: "excluir",
-                  onClick: () => excluirCertificado(record),
-                },
-              ],
-            }}
-          />
 
-        </Flex>
-        <Flex direction="column" className={style.funcionarios}>
-          <Flex justify="center" mb="xs">
-            <Title order={2} className={style.titulo}>
-              Funcionários
-            </Title>
-          </Flex>
-          <ListaFuncionarioJuridico id={id} />
-          <Flex
-            direction="row"
-            justify="flex-end"
-            mt="xs"
-            className={style.adicionarButton}
-          >
-            {/* <ModalCadFuncionarios id={id} /> */}
-          </Flex>
-        </Flex>
+        <Certificados id={id} />
+        
       </Flex>
-    </div>
+      <ListaFuncionarioJuridico id={id} />
+    </Flex>
   );
 }
